thepoetrypharmacy_com/requirements.txt
flask
sqlalchemy
progressbar2
python-dateutil